+++ src/main/java/lotr/client/render/tileentity/LOTRRenderUtumnoReturnPortal.java
@@ -1,13 +1,13 @@
 package lotr.client.render.tileentity;
 
-import org.lwjgl.opengl.GL11;
-
 import lotr.common.tileentity.LOTRTileEntityUtumnoReturnPortal;
 import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.tileentity.TileEntitySpecialRenderer;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.*;
+import net.minecraft.util.MathHelper;
+import net.minecraft.util.ResourceLocation;
 import net.minecraft.world.World;
+import org.lwjgl.opengl.GL11;
 
 public class LOTRRenderUtumnoReturnPortal extends TileEntitySpecialRenderer {
 	public static ResourceLocation lightCircle = new ResourceLocation("lotr:misc/utumnoPortal_lightCircle.png");
@@ -34,7 +34,7 @@ public class LOTRRenderUtumnoReturnPortal extends TileEntitySpecialRenderer {
 			GL11.glPushMatrix();
 			tessellator.startDrawingQuads();
 			tessellator.setColorRGBA_F(1.0f, 1.0f, 1.0f, alpha);
-			double width = (float) (i + 1) / (float) passes * 0.99f;
+			double width = (float) (i + 1) / passes * 0.99f;
 			double bottom = 0.0;
 			double top = bottom + LOTRTileEntityUtumnoReturnPortal.PORTAL_TOP;
 			tessellator.addVertexWithUV(width /= 2.0, bottom, width, 0.0, 0.0);
