+++ src/main/java/lotr/common/command/LOTRCommandTime.java
@@ -1,9 +1,11 @@
 package lotr.common.command;
 
-import java.util.List;
-
 import lotr.common.LOTRTime;
-import net.minecraft.command.*;
+import net.minecraft.command.CommandBase;
+import net.minecraft.command.ICommandSender;
+import net.minecraft.command.WrongUsageException;
+
+import java.util.List;
 
 public class LOTRCommandTime extends CommandBase {
 	@Override
@@ -37,7 +39,7 @@ public class LOTRCommandTime extends CommandBase {
 		if (args.length >= 2) {
 			if ("set".equals(args[0])) {
 				long time;
-				time = "day".equals(args[1]) ? Math.round(LOTRTime.DAY_LENGTH * 0.03) : "night".equals(args[1]) ? Math.round(LOTRTime.DAY_LENGTH * 0.6) : (long) CommandBase.parseIntWithMin(sender, args[1], 0);
+				time = "day".equals(args[1]) ? Math.round(LOTRTime.DAY_LENGTH * 0.03) : "night".equals(args[1]) ? Math.round(LOTRTime.DAY_LENGTH * 0.6) : CommandBase.parseIntWithMin(sender, args[1], 0);
 				LOTRTime.setWorldTime(time);
 				CommandBase.func_152373_a(sender, this, "commands.lotr.time.set", time);
 				return;
